+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.23.1 (2026-07-16)
+
+### Fix
+
+- **cloudflare**: improve checkbox click logic for faster failure handling
+- **cloudflare**: update checkbox selector and improve bypass logic
+- nested iframes
+- **mouse**: keep debug overlay visible across page navigations
+
 ## 2.23.0 (2026-05-22)
 
 ### Feat